Move raw component building to a separate file
Try to cut down on the size/complexity of `encoding.rs` by moving out
this piece to a dedicated builder file.
